% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Accumulation}
\alias{Accumulation}
\title{Accumulation hypothesis}
\usage{
Accumulation(...)
}
\arguments{
\item{...}{Binary exposure variables to accumulate.}
}
\value{
Derived accumulation variable, i.e. sum of exposures.
}
\description{
Function for deriving an accumulation variable
from longitudinally measured exposures.
}
