% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Always}
\alias{Always}
\title{Always hypothesis}
\usage{
Always(...)
}
\arguments{
\item{...}{Binary exposure variables to accumulate.}
}
\value{
Derived 'always' variable, i.e. set to
1 if exposure present for all exposure variables, and 0 otherwise.
}
\description{
Function for deriving an 'always' variable
from longitudinally measured exposures.
}
