% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Change}
\alias{Change}
\title{Change hypothesis}
\usage{
Change(x1, x2)
}
\arguments{
\item{x1}{First binary exposure measurement variable.}

\item{x2}{Second binary exposure measurement variable.}
}
\value{
Derived change variable, i.e. set to
1 for all cases where x1 differs from x2, otherwise 0.
}
\description{
Function for deriving a change variable
from longitudinally measured exposures.
}
