% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Ever}
\alias{Ever}
\title{Ever hypothesis}
\usage{
Ever(...)
}
\arguments{
\item{...}{Binary exposure variables to accumulate.}
}
\value{
Derived 'ever' variable, i.e.
set to 1 if exposure present for any exposure variable,
otherwise 0.
}
\description{
Function for deriving an 'ever' variable
from longitudinally measured exposures.
}
