% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{MI2dummy}
\alias{MI2dummy}
\title{Dummy datasets with specified means and covariances}
\usage{
MI2dummy(formula, data, seed = 1)
}
\arguments{
\item{formula}{An object of class `formula` which specifies the
SLCMA model including outcome variable, exposures
and lifecourse hypotheses.}

\item{data}{An optional data frame, list or environment
containing the variables in the model. If not found in `data`,
the variables are taken from `environment(formula)`, typically
the environment form which the function is called.}

\item{seed}{Integer input to `set.seed` to ensure that
dataset generation is reproducible (Default: 1).}
}
\value{
Data frame of class `MI2dummy` including randomly
generated variables specified by the specified `formula` with
means and covariances obtained from the input `data`.
}
\description{
Creates a dummy dataset with the same samples means
and sample covariances as that estimated from pooled imputed datasets.
}
