% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Mobility}
\alias{Mobility}
\title{Mobility hypothesis}
\usage{
Mobility(x1, x2)
}
\arguments{
\item{x1}{First binary exposure measurement variable.}

\item{x2}{Second binary exposure measurement variable.}
}
\value{
Derived mobility variables, i.e. matrix
of two columns, first column indicating
upward mobility (`(1-x1)*x2`) and the second
column indicating downward mobility (`(1-x2)*x1`).
}
\description{
Function for deriving a mobility variable
from longitudinally measured exposures.
}
