% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{plot.slcma}
\alias{plot.slcma}
\title{SLCMA elbow plot}
\usage{
\method{plot}{slcma}(
  x,
  relax = FALSE,
  show.remove = FALSE,
  show.labels = TRUE,
  labels = NULL,
  selection.order = FALSE,
  xlab = "Variables selected",
  ylab = ifelse(x$adjust.df > 0, "Partial R-squared", "R-squared"),
  ...
)
}
\arguments{
\item{x}{Object of class \code{slcma} from the \code{slcma()} function.}
}
\description{
Elbow plot (plot method for class "slcma")
}
