% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcma}
\alias{print.slcma}
\title{SLCMA stage 1 output}
\usage{
\method{print}{slcma}(x)
}
\arguments{
\item{x}{Object of class \code{slcma} from the \code{slcma()} function.}
}
\description{
Print method for objects of class "slcma".
}
