% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcmaBayes}
\alias{print.slcmaBayes}
\title{Print slcmaBayes}
\usage{
\method{print}{slcmaBayes}(x)
}
\arguments{
\item{x}{Object of class \code{slcmaBayes} from the \code{slcmaBayes} function.}
}
\description{
Print method for objects of class \code{slcmaBayes}
}
