% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcmaFLI}
\alias{print.slcmaFLI}
\title{Print slcmaFLI output}
\usage{
\method{print}{slcmaFLI}(x, tailarea = TRUE)
}
\arguments{
\item{x}{Object of class \code{slcmaFLI} from the \code{slcmaFLI} function.}

\item{tailarea}{Print tail areas of confidence intervals (Default: TRUE)}
}
\description{
Print method for class \code{slcmaFLI} (adapted from \code{selectiveinference} package).
}
