% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcmaInfer}
\alias{print.slcmaInfer}
\title{SLCMA stage 2 output}
\usage{
\method{print}{slcmaInfer}(x)
}
\arguments{
\item{x}{Object of class "`slcmaInfer`" from the `slcmaInfer()` function.}
}
\description{
Print method for objects of class `"slcmaInfer"`.
}
