% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcmaRelax}
\alias{print.slcmaRelax}
\title{Print slcmaRelax}
\usage{
\method{print}{slcmaRelax}(x)
}
\arguments{
\item{x}{Object of class \code{slcmaRelax} from the \code{slcmaRelax} function.}
}
\description{
Print method for objects of class \code{slcmaRelax}
}
