% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.summary.slcma}
\alias{print.summary.slcma}
\title{SLCMA summary output}
\usage{
\method{print}{summary.slcma}(x)
}
\arguments{
\item{x}{Object of class \code{summary.slcma} from the \code{summary.slcma()} function.}
}
\description{
Print method for objects of class \code{summary.slcma}.
}
