% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{Recency}
\alias{Recency}
\title{Recency hypothesis}
\usage{
Recency(weights, ...)
}
\arguments{
\item{weights}{Vector of numeric weights for combining exposures,
one per exposure.}

\item{...}{Binary exposure variables to combine.}
}
\value{
Derived recency variable, i.e. weighted
sum of exposures.
}
\description{
Function for deriving a recency variable
from longitudinally measured exposures.
}
