% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcma}
\alias{slcma}
\title{Structured Life Course Modelling Approach: stage 1}
\usage{
slcma(
  formula,
  data = environment(formula),
  adjust = NULL,
  seed = 1,
  silent = FALSE,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} which specifies the
SLCMA model including outcome variable, lifecourse hypotheses and exposures.}

\item{data}{An optional data frame, list or environment
containing the variables in the model. If not found in \code{data},
the variables are taken from \code{environment(formula)}, typically
the environment from which the function is called.}

\item{adjust}{A numerical vector containing the position of terms to
be adjusted for in all models. (Default: NULL - no adjustment).}

\item{seed}{Integer input to \code{set.seed} to ensure that dataset 
generation, if \code{data} contains multiple imputations, is reproducible (Default: 1).}

\item{silent}{Logical. The function does not print any output when TRUE (Default: FALSE).}

\item{...}{Additional arguments to be passed to \code{lars}.}
}
\value{
An list containing the following items:
\itemize{
\item \code{X_hypos} Numeric matrix with columns corresponding to SLCMA hypotheses.
\item \code{y_resid} Output variable, adjusted for terms specified in \code{adjust}.
\item \code{multiple} Output from multiple regression of all terms.
\item \code{sanity} Data frame containing details of all terms specified by \code{formula}
and whether they are included in all models.
\item \code{fit} Output of \code{lars}.
\item \code{base.r.squared} R-squared for the base model containing only those terms
specified in \code{adjust}.
\item \code{adjust.df} Degrees of freedom for the terms specified in \code{adjust}.
}
}
\description{
Performs stage 1 (variable selection) of the SLCMA
}
