% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcmaBayes}
\alias{slcmaBayes}
\title{Bayesian inference}
\usage{
slcmaBayes(x)
}
\arguments{
\item{x}{Object of class \code{slcma} from the \code{slcma()} function.}
}
\description{
Function for calculating the posterior probability that the first-selected variable is the most important
}
