% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcmaInfer}
\alias{slcmaInfer}
\title{Structured Approach to Evaluating Life-Course Hypotheses: stage 1}
\usage{
slcmaInfer(
  slcma,
  step = 1L,
  method = "slcmaFLI",
  alpha = 0.05,
  do.maxtCI = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional arguments to `slcmaFLI()`.}
}
\description{
Performs stage 2 of the SLCMA for user-specified methods
}
