% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcmaRelax}
\alias{slcmaRelax}
\title{Relaxed lasso inference}
\usage{
slcmaRelax(x, step, alpha = 0.05)
}
\arguments{
\item{x}{Object of class \code{slcma} from the \code{slcma()} function.}

\item{step}{Integer specifying which step of the LARS procedure produces
the model on which inference is to be performed.}

\item{alpha}{Level of significance for confidence interval calculations. Confidence
intervals will have \code{(1 - alpha) * 100}\% coverage. (Default: 0.05 - 95\% coverage)}
}
\description{
Function for simple (naive) confidence interval calculations
}
