% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{SpecifyingHypotheses}
\alias{SpecifyingHypotheses}
\alias{Accumulation}
\alias{Recency}
\alias{Change}
\alias{Mobility}
\alias{Always}
\alias{Ever}
\title{Specifying hypotheses}
\usage{
Accumulation(...)

Recency(weights, ...)

Change(...)

Mobility(...)

Always(...)

Ever(...)
}
\arguments{
\item{...}{Binary exposure variables.}

\item{weights}{Vector of numeric weights for combining exposures,
one per exposure.}
}
\value{
Derived hypothesis variable(s) as follows:
\itemize{
\item \code{Accumulation(...)} - sum of exposures
\item \code{Recency(weight,...)} - weighted sum of exposures
\item \code{Change(x1,x2)} - variable indicating where the two exposures differ
\item \code{Mobility(x1,x2)} - two variables with the first indicating upward mobility (i.e. \code{x1==0 && x2==1}) and the second indicating downward mobility (i.e. \code{x1==1 && x2==0})
\item \code{Always(...)} - variable indicating whether all exposures present
\item \code{Ever(...)} - variable indicating whether at least one exposure present
}
}
\description{
Functions for deriving variables corresponding
various hypotheses about longitudinally
measured exposures.
}
