% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{summary.slcma}
\alias{summary.slcma}
\title{SLCMA output summary}
\usage{
\method{summary}{slcma}(x)
}
\arguments{
\item{x}{Object of class "`slcma`" from the `slcma()` function.}
}
\description{
Summary method for class "slcma".
}
