% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{MI2dummy}
\alias{MI2dummy}
\title{Dummy datasets with specified means and covariances}
\usage{
MI2dummy(formula, data, seed = 1)
}
\arguments{
\item{formula}{An object of class \code{formula} that specifies all terms
to be included in the dataset.}

\item{data}{An object of class \code{mids} containing the variables
optional data frame, list or environment specified in \code{formula}.}

\item{seed}{Integer input to \code{set.seed} to ensure that
dataset generation is reproducible (Default: 1).}
}
\value{
Data frame of class \code{MI2dummy} including randomly
generated variables specified by the specified \code{formula} with
means and covariances obtained from the input \code{data}.
}
\description{
Creates a dummy dataset with the same samples means
and sample covariances as that estimated from pooled imputed datasets.
}
