% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{plot.slcma}
\alias{plot.slcma}
\title{SLCMA elbow plot}
\usage{
\method{plot}{slcma}(
  slcma,
  relax = FALSE,
  show.remove = FALSE,
  show.labels = TRUE,
  labels = NULL,
  selection.order = FALSE,
  xlab = "Variables selected",
  ylab = "R-squared",
  ...
)
}
\arguments{
\item{x}{Object of class "`slcma`" from the `slcma()` function.}
}
\description{
Elbow plot (plot method for class "slcma")
}
