% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{print.slcmaFLI}
\alias{print.slcmaFLI}
\title{Print slcmaFLI output}
\usage{
\method{print}{slcmaFLI}(x, tailarea = TRUE, ...)
}
\arguments{
\item{tailarea}{(Default: TRUE)}
}
\value{

}
\description{
Print method for class "slcmaFLI" (adapted from 'selectiveinference' package).
}
