% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcma}
\alias{slcma}
\title{Structured Approach to Evaluating Life-Course Hypotheses: stage 1}
\usage{
slcma(formula, data = environment(formula), adjust = NULL, ...)
}
\arguments{
\item{formula}{An object of class `formula` which specifies the
SLCMA model including outcome variable, exposures
and lifecourse hypotheses.}

\item{data}{An optional data frame, list or environment
containing the variables in the model. If not found in `data`,
the variables are taken from `environment(formula)`, typically
the environment form which the function is called.}

\item{adjust}{MISSING}

\item{...}{Additional arguments to be passed to `MI2dummy`.}
}
\value{
An object of class `"sclma"` which inherits from
the `"lars"` class containing output from the least angle regression
analysis as well as the following items:
* `X` Numeric matrix with columns corresponding to SLCMA hypotheses.
* `y` Output variable.
* `ncov`
* `multiple`
* `sanity`
}
\description{
Performs stage 1 of the SLCMA
}
