% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcma.R
\name{slcmaMaxt}
\alias{slcmaMaxt}
\title{Max-|t| test}
\usage{
slcmaMaxt(x, alpha = 0.05, do.CI = TRUE, seed = 12345, ...)
}
\arguments{
\item{alpha}{(Default: 0.05)}

\item{do.CI}{(Default: FALSE)}

\item{seed}{(Default: 12345)}
}
\description{
Function for max-|t| test and associated confidence intervals
}
